pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.0.1"
description = "Bot para manter a ordem no Kud"
authors = [
    "Leonardo Godoy <leonardoggavellar@gmail.com>",
    "Bernardo Lins <bernardolins28@gmail.com>"
]

[tool.poetry.dependencies]
python = ">=3.8.0,<4"
"discord.py" = "==2.0.1"
table2ascii = "^0.3.0"
